requirements_a5.txt
absl-py==1.3.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
audioread==3.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bibtexparser==1.2.0
blinker==1.5
bs4==0.0.1
cachetools==5.2.0
captum==0.5.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
cycler==0.11.0
datasets==2.4.0
decorator==5.1.1
dill==0.3.5.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.0
entrypoints==0.4
evaluate==0.4.0
ffmpeg-python==0.2.0
filelock==3.8.0
flatbuffers==22.12.6
fonttools==4.33.3
frozenlist==1.3.1
fsspec==2022.7.1
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.28
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.10.0
ipython==7.34.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
keras==2.11.0
kiwisolver==1.4.2
langdetect==1.0.9
libclang==14.0.6
librosa==0.9.2
llvmlite==0.38.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
multidict==6.0.2
multiprocess==0.70.13
nltk==3.8.1
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.4
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickle5==0.0.12
pickleshare==0.7.5
Pillow==9.1.0
pkgutil_resolve_name==1.3.10
pooch==1.6.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.9.13
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.4.0
responses==0.18.0
rich==12.6.0
rpunct
rsa==4.9
scikit-learn==0.24.2
scipy==1.6.3
semver==2.13.0
sentencepiece==0.1.97
sentry-sdk==1.9.10
seqeval==1.2.2
setproctitle==1.3.2
shortuuid==1.0.9
simpletransformers==0.61.4
six==1.16.0
sklearn==0.0
smmap==5.0.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
streamlit==1.13.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
threadpoolctl==3.1.0
timm==0.4.5
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
#torch==1.12.1+cu116
#torchaudio==0.12.1+cu116
#torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.64.0
traitlets==5.4.0
transformers==4.22.2
transformers-interpret==0.8.1
typing_extensions==4.2.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wandb==0.13.4
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
xxhash==3.0.0
yarl==1.8.1
zipp==3.4.1
